+++ svd-rs/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+## [v0.14.2] - 2023-04-04
+
 - Add support of `a-Z` for `dimIndex`
 - Add `name`, `description`, `address_offset` for `RegisterCluster`
 
@@ -79,7 +81,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Previous versions in common [changelog](../CHANGELOG.md).
 
-[Unreleased]: https://github.com/rust-embedded/svd/compare/svd-rs-v0.14.1...HEAD
+[Unreleased]: https://github.com/rust-embedded/svd/compare/svd-rs-v0.14.2...HEAD
+[v0.14.2]: https://github.com/rust-embedded/svd/compare/svd-encoder-v0.14.2...svd-rs-v0.14.2
 [v0.14.1]: https://github.com/rust-embedded/svd/compare/v0.14.0...svd-rs-v0.14.1
 [v0.14.0]: https://github.com/rust-embedded/svd/compare/svd-rs-v0.13.2...v0.14.0
 [v0.13.2]: https://github.com/rust-embedded/svd/compare/svd-rs-v0.13.1...svd-rs-v0.13.2